requirements.txt
comtypes~=1.1.14
pycaw~=20220416
Pillow~=9.3.0